ops/snapshot: separate generating and saving the versions
This commit separates the version generation from the package sets
from the one from packages (as most probably users will want to control
this). Moreover, it separates the work previously done by #commit into
the version generation (snapshot_*) and saving the new versions
(#versions)

The renaming of #versions into #snapshot_* is by consistency with
the Autobuild::Package API in which the method is called snapshot.