snapshot: make the overrides data a single array
The separation between 'version_control' and 'overrides' was
done in autoproj-snapshot because it was overwriting the overrides.yml
file. Since we don't have such problem with the files in overrides.d/,
we only need to define overrides (which greatly simplifies things).

In addition, it is now assumed that the files in overrides.d only
contains overrides information, and remove the toplevel hash. The
'overrides' field in autoproj/overrides.yml is still parsed as-is.

snapshot_package_sets now uses the package set's importer repository ID
(e.g. github:rock-core/package_set) as name.